doc: add sink elasticsearch README
