+++ connectors/sink-elasticsearch/internal/elasticsearch_sink.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sink_elasticsearch
+package internal
 
 import (
 	"bytes"
