+++ connectors/source-alicloud-billing/internal/alicloud_billing.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package alicloud_billing
+package internal
 
 import (
 	"context"
