+++ connectors/source-aws-billing/internal/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aws_billing
+package internal
 
 import (
 	"context"
